+++ TacticalAI/NPC.cpp
@@ -1891,7 +1891,7 @@ void ResetOncePerConvoRecordsForAllNPCsInLoadedSector( void )
 }
 
 
-void ReturnItemToPlayerIfNecessary( UINT8 ubMerc, INT8 bApproach, UINT32 uiApproachData, NPCQuoteInfo * pQuotePtr )
+void ReturnItemToPlayerIfNecessary( UINT8 ubMerc, INT8 bApproach, uintptr_t uiApproachData, NPCQuoteInfo * pQuotePtr )
 {
 	OBJECTTYPE  *		pObj;
 	SOLDIERTYPE *		pSoldier;
@@ -1914,7 +1914,7 @@ void ReturnItemToPlayerIfNecessary( UINT8 ubMerc, INT8 bApproach, UINT32 uiAppro
 	}
 }
 
-void Converse( UINT8 ubNPC, UINT8 ubMerc, INT8 bApproach, UINT32 uiApproachData )
+void Converse( UINT8 ubNPC, UINT8 ubMerc, INT8 bApproach, uintptr_t uiApproachData )
 {
 	NPCQuoteInfo					QuoteInfo;
 	NPCQuoteInfo *				pQuotePtr = &(QuoteInfo);
@@ -1927,7 +1927,7 @@ void Converse( UINT8 ubNPC, UINT8 ubMerc, INT8 bApproach, UINT32 uiApproachData 
 	SOLDIERTYPE *					pNPC;
 	BOOLEAN								fAttemptingToGiveItem;
 
-	DebugQuestInfo(String("Converse: merc <%d> to NPC <%d> approach %d data %d", ubMerc, ubNPC, bApproach, uiApproachData));
+	DebugQuestInfo(String("Converse: merc <%d> to NPC <%d> approach %d data %d", ubMerc, ubNPC, bApproach, (UINT32)uiApproachData));
 
 	// we have to record whether an item is being given in order to determine whether,
 	// in the case where the approach is overridden, we need to return the item to the
